Split logic into 2 scripts: 1. get_spring_boot_managed_external_dependencies.py 2. sync_external_dependencies.py
