+++ sdk/spring/scripts/get_spring_boot_managed_external_dependencies.py
@@ -1,37 +1,34 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
+############################################################################################################################################
+# This script is used to get all 3rd party dependencies managed by spring-boot-dependencies and spring-cloud-dependencies.
 #
-# How to use this script to update spring-boot's version?
-# Here are the steps:
-# 1. Update ROOT_POMS' version manually.
-# 2. Run command "python .\sdk\spring\scripts\update_spring_managed_external_dependencies.py".
-# 3. Run command "python .\eng\versioning\update_versions.py --ut external_dependency --sr".
-# 4. Run command ".\eng\versioning\pom_file_version_scanner.ps1 -Debug". If there is error, fix it.
-# 5. Update changelog about compatible Spring Boot versions and Spring Cloud versions.
+# How to use this script.
+#  1. Update `SPRING_BOOT_VERSION` and `SPRING_CLOUD_VERSION` in this script manually.
+#  2. Run command `python .\sdk\spring\scripts\get_spring_boot_managed_external_dependencies.py`.
+#  3. Then a file named `spring_boot_SPRING_BOOT_VERSION_managed_external_dependencies.txt` will be created.
+#
+# Please refer to ./README.md to get more information about this script.
+############################################################################################################################################
 
 import argparse
 import os
 import queue
 import time
-import unittest
 import urllib.request as request
 import xml.etree.ElementTree as elementTree
-from itertools import takewhile
 from urllib.error import HTTPError
-import in_place
 
 from log import log, Log
 from pom import Pom
 
-EXTERNAL_DEPENDENCIES_FILE = 'eng/versioning/external_dependencies.txt'
+SPRING_BOOT_VERSION = '2.5.4'
+SPRING_CLOUD_VERSION = '2020.0.3'
+
 ROOT_POMS = [
-    'org.springframework.boot:spring-boot-starter-parent;2.5.4',
-    'org.springframework.boot:spring-boot-dependencies;2.5.4',
-    'org.springframework.cloud:spring-cloud-dependencies;2020.0.3'
+    'org.springframework.boot:spring-boot-starter-parent;{}'.format(SPRING_BOOT_VERSION),
+    'org.springframework.boot:spring-boot-dependencies;{}'.format(SPRING_BOOT_VERSION),
+    'org.springframework.cloud:spring-cloud-dependencies;{}'.format(SPRING_CLOUD_VERSION)
 ]
-SKIP_IDS = [
-    'org.eclipse.jgit:org.eclipse.jgit'  # Refs: https://github.com/Azure/azure-sdk-for-java/pull/13956/files#r468368271
-]
+SPRING_BOOT_MANAGED_EXTERNAL_DEPENDENCIES_FILE_NAME = 'sdk/spring/scripts/spring_boot_{}_managed_external_dependencies.txt'.format(SPRING_BOOT_VERSION)
 MAVEN_NAME_SPACE = {'maven': 'http://maven.apache.org/POM/4.0.0'}
 
 
@@ -42,8 +39,7 @@ def main():
     dependency_dict = {}
     for root_pom in ROOT_POMS:
         update_dependency_dict(dependency_dict, root_pom)
-    output_version_dict_to_file(dependency_dict)
-    update_version_for_external_dependencies(dependency_dict)
+    output_version_dict_to_file(dependency_dict, SPRING_BOOT_MANAGED_EXTERNAL_DEPENDENCIES_FILE_NAME)
     elapsed_time = time.time() - start_time
     log.info('elapsed_time = {}'.format(elapsed_time))
 
@@ -76,7 +72,7 @@ def update_dependency_dict(dependency_dict, root_pom_id):
         pom_url = pom.to_url()
         log.info('Get dependencies from pom. depth = {}, url = {}.'.format(pom.depth, pom_url))
         try:
-            tree = elementTree.ElementTree(file = request.urlopen(pom_url))
+            tree = elementTree.ElementTree(file=request.urlopen(pom_url))
         except HTTPError:
             log.warn('Error in open {}'.format(pom_url))
             continue
@@ -90,7 +86,7 @@ def update_dependency_dict(dependency_dict, root_pom_id):
             parent_version = parent_element.find('./maven:version', MAVEN_NAME_SPACE).text.strip(' ${}')
             parent_pom = Pom(parent_group_id, parent_artifact_id, parent_version, pom.depth + 1)
             parent_pom_url = parent_pom.to_url()
-            parent_tree = elementTree.ElementTree(file = request.urlopen(parent_pom_url))
+            parent_tree = elementTree.ElementTree(file=request.urlopen(parent_pom_url))
             parent_project_element = parent_tree.getroot()
             log.debug('Get properties from parent pom. parent_pom_url = {}.'.format(parent_pom_url))
             update_property_dict(parent_project_element, property_dict)
@@ -114,10 +110,8 @@ def update_dependency_dict(dependency_dict, root_pom_id):
                 log.info('Dependency version skipped. key = {}, version = {}, dependency_dict[key] = {}.'.format(key, version, dependency_dict[key]))
             artifact_type = dependency_element.find('./maven:type', MAVEN_NAME_SPACE)
             artifact_scope = dependency_element.find('./maven:scope', MAVEN_NAME_SPACE)
-            if artifact_type is not None and \
-                artifact_scope is not None and \
-                artifact_type.text.strip() == 'pom' and \
-                artifact_scope.text.strip() == 'import':
+            if artifact_type is not None and artifact_scope is not None and \
+                    artifact_type.text.strip() == 'pom' and artifact_scope.text.strip() == 'import':
                 new_pom = Pom(group_id, artifact_id, version, pom.depth + 1)
                 q.put(new_pom)
                 pom_count = pom_count + 1
@@ -156,55 +150,11 @@ def update_property_dict(project_element, property_dict):
     return property_dict
 
 
-def output_version_dict_to_file(dependency_dict):
-    output_file = open('sdk/spring/scripts/spring_managed_external_dependencies.txt', 'w''')
+def output_version_dict_to_file(dependency_dict, output_file):
+    output_file = open(output_file, 'w''')
     for key, value in sorted(dependency_dict.items()):
         output_file.write('{};{}\n'.format(key, value))
     output_file.close()
-
-
-def update_version_for_external_dependencies(dependency_dict):
-    with in_place.InPlace(EXTERNAL_DEPENDENCIES_FILE) as file:
-        for line in file:
-            line = line.strip()
-            if line.startswith('#') or not line:
-                file.write(line)
-            else:
-                key_value = line.split(';', 1)
-                key = key_value[0]
-                value = key_value[1]
-                if key not in SKIP_IDS and key in dependency_dict:
-                    value_in_dict = dependency_dict[key]
-                    if version_bigger_than(value, value_in_dict):
-                        log.warn('Version update skipped. key = {}, value = {}, new_value = {}'.format(key, value, value_in_dict))
-                        file.write(line)
-                    elif version_bigger_than(value_in_dict, value):
-                        log.info('Version updated. key = {}, value = {}, new_value = {}'.format(key, value, value_in_dict))
-                        file.write('{};{}'.format(key, value_in_dict))
-                    else:
-                        file.write(line)
-                else:
-                    file.write(line)
-            file.write('\n')
-
-
-def version_bigger_than(version1, version2):
-    v1 = version1.split('.')
-    v2 = version2.split('.')
-    len_1 = len(v1)
-    len_2 = len(v2)
-    max_len = max(len_1, len_1)
-    for i in range(max_len):
-        if i < len_1 and i < len_2:
-            int_1 = int('0' + ''.join(takewhile(str.isdigit, v1[i])))
-            int_2 = int('0' + ''.join(takewhile(str.isdigit, v2[i])))
-            if int_1 != int_2:
-                return int_1 > int_2
-        elif i < len_1:
-            return True
-        else:
-            return False
-    return False
 
 
 def print_dict(d):
@@ -212,33 +162,17 @@ def print_dict(d):
         print('key = {}, value = {}.'.format(key, value))
 
 
-class Tests(unittest.TestCase):
-    def test_version_bigger_than(self):
-        self.assertEqual(version_bigger_than('1', '2'), False)
-        self.assertEqual(version_bigger_than('2', '1'), True)
-        self.assertEqual(version_bigger_than('1.0', '2'), False)
-        self.assertEqual(version_bigger_than('2.0', '1'), True)
-        self.assertEqual(version_bigger_than('1.1', '1'), True)
-        self.assertEqual(version_bigger_than('1', '1.1'), False)
-        self.assertEqual(version_bigger_than('1.0-RELEASE', '1.1'), False)
-        self.assertEqual(version_bigger_than('1.1-RELEASE', '1'), True)
-        self.assertEqual(version_bigger_than('1.1-RELEASE', '1.0'), True)
-        self.assertEqual(version_bigger_than('1.1-RELEASE', '1.0.1'), True)
-        self.assertEqual(version_bigger_than('1.1-RELEASE', '1.0.1-RELEASE'), True)
-        self.assertEqual(version_bigger_than('1.1-RELEASE', '1.1.1-RELEASE'), False)
-
-
 def init():
     parser = argparse.ArgumentParser(
-        description = 'Update versions in /eng/versioning/external_dependencies.txt.'
+        description='Get spring-boot managed external dependencies and write into {}'.format(SPRING_BOOT_MANAGED_EXTERNAL_DEPENDENCIES_FILE_NAME)
     )
     parser.add_argument(
         '--log',
-        type = str,
-        choices = ['debug', 'info', 'warn', 'error', 'none'],
-        required = False,
-        default = 'info',
-        help = 'Set log level.'
+        type=str,
+        choices=['debug', 'info', 'warn', 'error', 'none'],
+        required=False,
+        default='info',
+        help='Set log level.'
     )
     args = parser.parse_args()
     log_dict = {
