+++ sdk/spring/scripts/spring_boot_2.5.4_managed_external_dependencies.txt
@@ -81,7 +81,7 @@ com.github.tomakehurst:wiremock-jre8;2.26.3
 com.github.tomakehurst:wiremock-jre8-standalone;2.27.1
 com.google.appengine:appengine-api-1.0-sdk;1.9.91
 com.google.cloud:cloud-spanner-r2dbc;0.5.0
-com.google.code.gson:gson;2.8.9
+com.google.code.gson:gson;2.8.7
 com.h2database:h2;1.4.200
 com.hazelcast:hazelcast;4.1.5
 com.hazelcast:hazelcast-hibernate52;2.2.1
@@ -1353,28 +1353,28 @@ org.springframework.restdocs:spring-restdocs-restassured;2.0.5.RELEASE
 org.springframework.restdocs:spring-restdocs-webtestclient;2.0.5.RELEASE
 org.springframework.retry:spring-retry;1.3.1
 org.springframework.security.oauth.boot:spring-security-oauth2-autoconfigure;2.1.2.RELEASE
-org.springframework.security:spring-security-acl;5.6.0
-org.springframework.security:spring-security-aspects;5.6.0
-org.springframework.security:spring-security-bom;5.6.0
-org.springframework.security:spring-security-cas;5.6.0
-org.springframework.security:spring-security-config;5.6.0
-org.springframework.security:spring-security-core;5.6.0
-org.springframework.security:spring-security-crypto;5.6.0
-org.springframework.security:spring-security-data;5.6.0
-org.springframework.security:spring-security-ldap;5.6.0
-org.springframework.security:spring-security-messaging;5.6.0
-org.springframework.security:spring-security-oauth2-client;5.6.0
-org.springframework.security:spring-security-oauth2-core;5.6.0
-org.springframework.security:spring-security-oauth2-jose;5.6.0
-org.springframework.security:spring-security-oauth2-resource-server;5.6.0
-org.springframework.security:spring-security-openid;5.6.0
-org.springframework.security:spring-security-remoting;5.6.0
+org.springframework.security:spring-security-acl;5.5.2
+org.springframework.security:spring-security-aspects;5.5.2
+org.springframework.security:spring-security-bom;5.5.2
+org.springframework.security:spring-security-cas;5.5.2
+org.springframework.security:spring-security-config;5.5.2
+org.springframework.security:spring-security-core;5.5.2
+org.springframework.security:spring-security-crypto;5.5.2
+org.springframework.security:spring-security-data;5.5.2
+org.springframework.security:spring-security-ldap;5.5.2
+org.springframework.security:spring-security-messaging;5.5.2
+org.springframework.security:spring-security-oauth2-client;5.5.2
+org.springframework.security:spring-security-oauth2-core;5.5.2
+org.springframework.security:spring-security-oauth2-jose;5.5.2
+org.springframework.security:spring-security-oauth2-resource-server;5.5.2
+org.springframework.security:spring-security-openid;5.5.2
+org.springframework.security:spring-security-remoting;5.5.2
 org.springframework.security:spring-security-rsa;1.0.10.RELEASE
-org.springframework.security:spring-security-rsocket;5.6.0
-org.springframework.security:spring-security-saml2-service-provider;5.6.0
-org.springframework.security:spring-security-taglibs;5.6.0
-org.springframework.security:spring-security-test;5.6.0
-org.springframework.security:spring-security-web;5.6.0
+org.springframework.security:spring-security-rsocket;5.5.2
+org.springframework.security:spring-security-saml2-service-provider;5.5.2
+org.springframework.security:spring-security-taglibs;5.5.2
+org.springframework.security:spring-security-test;5.5.2
+org.springframework.security:spring-security-web;5.5.2
 org.springframework.session:spring-session-bom;2021.0.2
 org.springframework.session:spring-session-core;2.5.2
 org.springframework.session:spring-session-data-geode;2.5.4
